requirements.txt
python-telegram-bot
pyyaml
pytz
arrow
ics
requests